refactor: 가게 삭제 API에 사용자 인증 로직 추가(본인 가게만 삭제할 수 있도록 수정)

+++ src/main/java/com/example/BE/store/service/StoreService.java
@@ -8,6 +8,7 @@ import com.example.BE.store.domain.dto.StoreDetailResponse;
 import com.example.BE.store.domain.dto.StoreResponse;
 import com.example.BE.store.domain.dto.StoreUpdateRequest;
 import com.example.BE.store.repository.StoreRepository;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -73,8 +74,13 @@ public class StoreService {
     }
 
     @Transactional
-    public void deleteStore(Long storeId) {
-        findByIdOrThrow(storeId);
+    public void deleteStore(Long userId, Long storeId) {
+        StoreEntity store = findByIdOrThrow(storeId);
+
+        if(!store.getUserId().equals(userId)){
+            throw CustomException.from(StoreErrorCode.FORBIDDEN_STORE_ACCESS);
+        }
+
         storeRepository.deleteById(storeId);
     }
